+++ examples/react/next-server-actions/src/app/shared-code.ts
@@ -1,4 +1,4 @@
-import { formOptions } from '@tanstack/react-form/nextjs'
+import { formOptions } from '@tanstack/react-form-nextjs'
 
 export const formOpts = formOptions({
   defaultValues: {