+++ packages/form-core/src/utils.ts
@@ -1,7 +1,8 @@
-import { defaultValidationLogic } from './ValidationLogic'
+import { liteThrottle } from '@tanstack/pacer-lite'
+import { formEventClient } from './EventClient'
 import type { ValidationLogicProps } from './ValidationLogic'
 import type { FieldValidators } from './FieldApi'
-import type { FormValidators } from './FormApi'
+import type { AnyFormApi, FormValidators } from './FormApi'
 import type {
   GlobalFormValidationError,
   ValidationCause,
@@ -31,9 +32,9 @@ export function functionalUpdate<TInput, TOutput = TInput>(
  * Get a value from an object using a path, including dot notation.
  * @private
  */
-export function getBy(obj: any, path: any) {
+export function getBy(obj: unknown, path: string | (string | number)[]): any {
   const pathObj = makePathArray(path)
-  return pathObj.reduce((current: any, pathPart: any) => {
+  return pathObj.reduce((current: any, pathPart) => {
     if (current === null) return null
     if (typeof current !== 'undefined') {
       return current[pathPart]
@@ -108,7 +109,10 @@ export function deleteBy(obj: any, _path: any) {
 
     const key = path.shift()
 
-    if (typeof key === 'string') {
+    if (
+      typeof key === 'string' ||
+      (typeof key === 'number' && !Array.isArray(parent))
+    ) {
       if (typeof parent === 'object') {
         return {
           ...parent,
@@ -600,3 +604,14 @@ export function uuid(): string {
   IDX++
   return out
 }
+
+export const throttleFormState = liteThrottle(
+  (form: AnyFormApi) =>
+    formEventClient.emit('form-state', {
+      id: form.formId,
+      state: form.store.state,
+    }),
+  {
+    wait: 300,
+  },
+)