+++ packages/react-form-remix/vite.config.ts
@@ -1,0 +1,26 @@
+import { defineConfig, mergeConfig } from 'vitest/config'
+import { tanstackViteConfig } from '@tanstack/vite-config'
+import react from '@vitejs/plugin-react'
+import packageJson from './package.json'
+
+const config = defineConfig({
+  plugins: [react()],
+  test: {
+    name: packageJson.name,
+    dir: './tests',
+    watch: false,
+    environment: 'jsdom',
+    setupFiles: ['./tests/test-setup.ts'],
+    coverage: { enabled: true, provider: 'istanbul', include: ['src/**/*'] },
+    typecheck: { enabled: true },
+  },
+})
+
+export default mergeConfig(
+  config,
+  tanstackViteConfig({
+    entry: ['./src/index.ts'],
+    cjs: false,
+    srcDir: './src',
+  }),
+)