+++ packages/react-form/src/createFormHook.tsx
@@ -1,3 +1,4 @@
+'use client'
 /* eslint-disable @eslint-react/no-context-provider */
 import { createContext, useContext, useMemo } from 'react'
 import { useForm } from './useForm'
@@ -16,8 +17,8 @@ import type {
 import type {
   ComponentType,
   Context,
+  FunctionComponent,
   PropsWithChildren,
-  ReactNode,
 } from 'react'
 import type { FieldComponent } from './useField'
 import type { ReactFormExtendedApi } from './useForm'
@@ -190,7 +191,10 @@ export type AppFieldExtendedReactFormApi<
       TSubmitMeta,
       NoInfer<TFieldComponents>
     >
-    AppForm: ComponentType<PropsWithChildren>
+    AppForm: ComponentType<
+      // PropsWithChildren<P> is not optional in React 17
+      PropsWithChildren<{}>
+    >
   }
 
 export interface WithFormProps<
@@ -210,23 +214,23 @@ export interface WithFormProps<
   TFormComponents extends Record<string, ComponentType<any>>,
   TRenderProps extends object = Record<string, never>,
 > extends FormOptions<
-    TFormData,
-    TOnMount,
-    TOnChange,
-    TOnChangeAsync,
-    TOnBlur,
-    TOnBlurAsync,
-    TOnSubmit,
-    TOnSubmitAsync,
-    TOnDynamic,
-    TOnDynamicAsync,
-    TOnServer,
-    TSubmitMeta
-  > {
+  TFormData,
+  TOnMount,
+  TOnChange,
+  TOnChangeAsync,
+  TOnBlur,
+  TOnBlurAsync,
+  TOnSubmit,
+  TOnSubmitAsync,
+  TOnDynamic,
+  TOnDynamicAsync,
+  TOnServer,
+  TSubmitMeta
+> {
   // Optional, but adds props to the `render` function outside of `form`
   props?: TRenderProps
-  render: (
-    props: PropsWithChildren<
+  render: FunctionComponent<
+    PropsWithChildren<
       NoInfer<TRenderProps> & {
         form: AppFieldExtendedReactFormApi<
           TFormData,
@@ -245,8 +249,8 @@ export interface WithFormProps<
           TFormComponents
         >
       }
-    >,
-  ) => ReactNode
+    >
+  >
 }
 
 export interface WithFieldGroupProps<
@@ -258,8 +262,8 @@ export interface WithFieldGroupProps<
 > extends BaseFormOptions<TFieldGroupData, TSubmitMeta> {
   // Optional, but adds props to the `render` function outside of `form`
   props?: TRenderProps
-  render: (
-    props: PropsWithChildren<
+  render: FunctionComponent<
+    PropsWithChildren<
       NoInfer<TRenderProps> & {
         group: AppFieldExtendedReactFieldGroupApi<
           unknown,
@@ -282,8 +286,8 @@ export interface WithFieldGroupProps<
           TFormComponents
         >
       }
-    >,
-  ) => ReactNode
+    >
+  >
 }
 
 export function createFormHook<
@@ -341,13 +345,13 @@ export function createFormHook<
   > {
     const form = useForm(props)
 
-    const AppForm = useMemo(() => {
-      const AppForm = (({ children }) => {
+    // PropsWithChildren<P> is not optional in React 17
+    const AppForm = useMemo<ComponentType<PropsWithChildren<{}>>>(() => {
+      return ({ children }) => {
         return (
           <formContext.Provider value={form}>{children}</formContext.Provider>
         )
-      }) as ComponentType<PropsWithChildren>
-      return AppForm
+      }
     }, [form])
 
     const AppField = useMemo(() => {
@@ -520,7 +524,7 @@ export function createFormHook<
         fields: TFields
       }
     >,
-  ) => ReactNode {
+  ) => ReturnType<FunctionComponent> {
     return function Render(innerProps) {
       const fieldGroupProps = useMemo(() => {
         return {