+++ packages/form-core/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @tanstack/form-core
 
+## 1.27.4
+
+### Patch Changes
+
+- fix: prevent unnecessary re-renders when there are no async validators ([#1929](https://github.com/TanStack/form/pull/1929))
+
+  Fields were re-rendering twice on each keystroke because `isValidating` was being set to `true` then `false` even when there were no async validators to run. This fix checks if there are actual async validators before toggling the `isValidating` state.
+
+  Fixes #1130
+
+## 1.27.3
+
+### Patch Changes
+
+- Bump TanStack pacer to pacer-lite for reduced custom event emissions. ([#1876](https://github.com/TanStack/form/pull/1876))
+
+## 1.27.2
+
+## 1.27.1
+
+### Patch Changes
+
+- Fix issues with methods not being present in React adapter ([#1903](https://github.com/TanStack/form/pull/1903))
+
+## 1.27.0
+
+### Patch Changes
+
+- Fixed issues with React Compiler ([#1893](https://github.com/TanStack/form/pull/1893))
+
+- Fix issue with deleteField and numeric keys ([#1891](https://github.com/TanStack/form/pull/1891))
+
+## 1.26.0
+
+### Patch Changes
+
+- fix stale fields on array changes ([#1729](https://github.com/TanStack/form/pull/1729))
+- allow explicitly setting `field.handleChange(undefined)` ([#1729](https://github.com/TanStack/form/pull/1729))
+
+## 1.25.0
+
+### Patch Changes
+
+- Removes debug config from event client in form-core ([#1852](https://github.com/TanStack/form/pull/1852))
+
+## 1.24.5
+
+### Patch Changes
+
+- - Make `fieldMeta` record type `Partial<>` to reflect runtime behaviour ([#1787](https://github.com/TanStack/form/pull/1787))
+
 ## 1.24.4
 
 ### Patch Changes